+++ decidim-core/spec/system/homepage_spec.rb
@@ -28,6 +28,30 @@ describe "Homepage", type: :system do
       expect(page).to have_selector("a.main-footer__badge[href='#{official_url}']")
     end
 
+    context "and the organization has the omnipresent banner enabled" do
+      let(:organization) do
+        create(:organization,
+               official_url: official_url,
+               enable_omnipresent_banner: true,
+               omnipresent_banner_url: "#{official_url}/processes",
+               omnipresent_banner_title: Decidim::Faker::Localized.sentence(3),
+               omnipresent_banner_short_description: Decidim::Faker::Localized.sentence(3))
+      end
+
+      before do
+        switch_to_host(organization.host)
+        visit decidim.root_path
+      end
+
+      it "shows the omnipresent banner's title" do
+        expect(page).to have_i18n_content(organization.omnipresent_banner_title)
+      end
+
+      it "shows the omnipresent banner's short description" do
+        expect(page).to have_i18n_content(organization.omnipresent_banner_short_description)
+      end
+    end
+
     describe "call to action" do
       let!(:participatory_process) { create :participatory_process, :published }
       let!(:organization) { participatory_process.organization }