+++ CHANGELOG.md
@@ -13,6 +13,10 @@ you need to change the following line in `config/environments/production.rb`:
 ```
 
 **Added
+- **decidim-core**: Added support for omnipresent banner.
++[\#2547](https://github.com/decidim/decidim/pull/2547)
+- **decidim-admin**: Added administration interface for omnipresent banner.
++[\#2547](https://github.com/decidim/decidim/pull/2547)
 - **decidim**: Rubocop has been upgraded. Ruby version in gem specs now fits with ruby version that rubocop uses.
 [\#2548](https://github.com/decidim/decidim/pull/2548)
 - **decidim-admin**: Notify participatory process followers when a step becomes active. [\#2544](https://github.com/decidim/decidim/pull/2544)